+++ ec-prueba-unexus-servicio-tony-veas/src/main/java/ec/unexus/prueba/servicio/tony/veas/EcPruebaUnexusServicioTonyVeasApplication.java
@@ -1,7 +1,8 @@
-package ec.unexus.prueba.servicio.tony.veas.application;
+package ec.unexus.prueba.servicio.tony.veas;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
 
 /*
  * Es el punto de entrada principal para una aplicación Spring Boot
