+++ ec-prueba-unexus-servicio-tony-veas/src/test/java/ec/unexus/prueba/servicio/tony/veas/EcPruebaUnexusServicioTonyVeasApplicationTests.java
@@ -1,4 +1,4 @@
-package ec.unexus.prueba.servicio.tony.veas.application;
+package ec.unexus.prueba.servicio.tony.veas;
 
 import static org.assertj.core.api.Assertions.assertThat;
